tests-skeletons/check-APER-support: Showcase bug in aper_put_length()
This commit adds a test case which shows a bug introduced in
0101252ec65e3257fc19daeff26298c92531bc33 [1].

As a result, a small value (under 1 octet) with range spanning full 2
octets (65536) is encoded incorrectly as 1 octet.

Section 4 of X.691 defines 64K as 65536, and Section 11.5.7 explicitly states
"""
c) "range" is greater than 256 and less than or equal to 64K (the two-octet case).
"""

[1] https://github.com/mouse07410/asn1c/pull/91